requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
transformers==2.8.0
pandas
tqdm
pytorch-lightning
sklearn
ipdb
